requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-Pydantic==0.12.0
Flask-SQLAlchemy==3.1.1
gotrue==2.6.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
postgrest==0.16.9
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
PySocks==1.7.1
pytest==8.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
storage3==0.7.7
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
tomli==2.0.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
